commit 02 - 05/07

+++ game_functions.py
@@ -1,7 +1,7 @@
 from random import randint
 
 def damage(base, intensity):
-    min_d = 5  * intensity
-    max_d = 25 * intensity
-    atack = randint(((base * min_d) / 100), (base * max_d) / 100)
-    return int(atack)
+    min_d = int(5  * intensity)
+    max_d = int(25 * intensity)
+    atack = randint((int((base * min_d) / 100)), int((base * max_d) / 100))
+    return atack
